주문 생성 테스트 - 패키지 구조 REFACTORING
- 내가 잘못 넣었던 거 수정
